Some renaming and commenting models
