+++ TODO.md
@@ -1,23 +1,35 @@
-# Models
-
-## Forecasting
+# Forecasting
 - seq2seq (https://www.youtube.com/watch?v=VYpAodcdFfA)
 
-## Regression
-- Gaussian process (http://www.gaussianprocess.org/gpml/chapters/RW2.pdf)
+# Regression
+- Gaussian processes (http://www.gaussianprocess.org/gpml/chapters/RW2.pdf)
 - Quantile regression forest
 
 # Backtesting
 - Parameterize the moving window
 - Track weights history in `backtest_regression()`
 
-# Useful TimeSeries
+# TimeSeries
+- Multivariate TimeSeries
+- Support categorical time series
 
+# TimeSeries Generation
 The following could be useful TimeSeries:
+- Time of day, day of week, week of month, month of year, etc
 - Indicator variable for holidays (1 during holidays 0 else) in many regions
 - Indicator for quarters
 
+# Metrics
+- Correlation coefficient
+
 # Utils
+- More preprocessing: moving average, filters, ...
 - A function such as `checkresiduals()` used here https://otexts.com/fpp2/regression-evaluation.html
 
+# Vizualisation
+- seasonal plots (incl polar)
+- scatterplots (for multivariate) and scatterplot matrix
+
 # Others
+- Collection of datasets
+- Hierarchical reconciliation
