+++ u8timeseries/tests/test_statistics.py
@@ -3,7 +3,7 @@ import pandas as pd
 import numpy as np
 
 from u8timeseries import TimeSeries
-from u8timeseries.models.statistics import check_seasonality
+from u8timeseries.utils.statistics import check_seasonality
 
 
 class TimeSeriesTestCase(unittest.TestCase):
